+++ SerbleChat.Backend/Database/Repos/IFriendshipRepo.cs
@@ -6,6 +6,6 @@ public interface IFriendshipRepo {
     Task<Friendship[]> GetFriendships(string userId);
     Task<Friendship?> GetFriendship(string user1Id, string user2Id);
     Task AddFriendship(Friendship friendship);
-    Task RemoveFriendship(int id);
+    Task RemoveFriendship(long id);
     Task ModifyFriendship(Friendship friendship);
 }
