+++ SerbleChat.Backend/Database/Structs/ChannelRead.cs
@@ -5,15 +5,15 @@ namespace SerbleChat.Backend.Database.Structs;
 
 public class ChannelRead {
     [Key]
-    public int Id { get; set; }
+    public long Id { get; set; }
     
     [ForeignKey(nameof(ChannelNavigation))]
-    public int ChannelId { get; set; }
+    public long ChannelId { get; set; }
     
     [StringLength(64), ForeignKey(nameof(UserNavigation))]
     public string UserId { get; set; } = null!;
     
-    public int LastReadMessageId { get; set; }
+    public long LastReadMessageId { get; set; }
     
     // Navigation Properties
     public Channel ChannelNavigation { get; set; } = null!;
